homepage fixes
