+++ src/components/Navbar/About/AboutHero.tsx
