+++ src/components/Navbar/About/AboutText.tsx
