+++ src/components/Navbar/Product/ProductText.tsx
@@ -20,21 +20,21 @@ const ProductText = () => {
           </Box>
 
           <Button
-                    my={10}
-                    bg="brand.primary700"
-                    color="white"
-                    _hover={{ bg: "brand.primary600" }}
-                    rounded="full"
-                    as="a"
-                    href={MERCHANT_APP_URL}
-                    target="_blank"
-                    size={{ base: "sm", md: "md", lg: "lg" }}
-                    px={{ base: 4, md: 6, lg: 8 }}
-                    py={{ base: 2, md: 3, lg: 4 }}
-                  >
-                    <Box marginRight="10px" fontSize={{ base: "sm", md: "md", lg: "lg" }}>Get Started With One Funnel</Box>
-                    <ArrowRight size={iconSize} variant="TwoTone" color="white" />
-                  </Button>
+            my={10}
+            bg="brand.primary700"
+            color="white"
+            _hover={{ bg: "brand.primary600" }}
+            rounded="full"
+            as="a"
+            href={MERCHANT_APP_URL}
+            target="_blank"
+            size={{ base: "sm", md: "md", lg: "lg" }}
+            px={{ base: 4, md: 6, lg: 8 }}
+            py={{ base: 2, md: 3, lg: 4 }}
+          >
+            <Box marginRight="10px" fontSize={{ base: "sm", md: "md", lg: "lg" }}>Get Started With One Funnel</Box>
+            <ArrowRight size={iconSize} variant="TwoTone" color="white" />
+          </Button>
         </Box>
       </Box>
     </>
